build: move extension build into pyproject
